config/asgi.py
"""
ASGI config for Age of Voyage project.

It exposes the ASGI callable as a module-level variable named ``application``.
This is also configured for WebSockets with Channels.
"""

import os
from django.core.asgi import get_asgi_application
from channels.routing import ProtocolTypeRouter, URLRouter
from channels.auth import AuthMiddlewareStack

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')

django_asgi_app = get_asgi_application()

application = ProtocolTypeRouter({
    "http": django_asgi_app,
    "websocket": AuthMiddlewareStack(
        URLRouter([
            # WebSocket URL patterns will be added here
        ])
    ),
})
